Melhorias

+++ Backend/API/config/app/infra/eventosBanco.js
@@ -66,9 +66,9 @@ module.exports =  function(){
     }
 
     this.remover = function(connection, objeto, callback){
-
-        fs.unlinkSync(objeto.img);
-
+         
+        if(objeto.img != null) fs.unlinkSync(objeto.img);
+        
         var string = "DELETE FROM produto WHERE id = " + objeto.id;
 
         connection.query(string, callback);
